Reorganized Code
