--- src/ITMID.java
@@ -1,3 +1,0 @@
-public enum ITMID {
-    Bandana();
-}
