+++ src/MANIFEST.MF
@@ -1,0 +1,4 @@
+Manifest-Version: 1.0
+Created-By: 1.8.0_92 (Oracle Corporation)
+Main-Class: MrSSS.Game
+
