+++ src/MrSSS/IO/GUI/Window.java
@@ -1,0 +1,26 @@
+package MrSSS.IO.GUI;
+
+import java.awt.Canvas;
+import java.awt.Dimension;
+
+import javax.swing.JFrame;
+
+import MrSSS.Game;
+
+public class Window extends Canvas {
+    private static final long serialVersionUID = -240840600533728354L;
+
+    public Window(int width, int height, String title, Game game) {
+		JFrame frame = new JFrame(title);
+		
+		frame.setPreferredSize(new Dimension(width, height));
+		frame.setMinimumSize(new Dimension(width, height));
+
+		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+		frame.setResizable(false);
+		frame.setLocationRelativeTo(null);
+		frame.add(game);
+		frame.setVisible(true);
+		game.start();
+    }
+}
