+++ src/MrSSS/IO/Loader/JarLoader.java
@@ -1,0 +1,3 @@
+package MrSSS.IO.Loader;
+
+public class JarLoader {}
