+++ src/MrSSS/IO/Loader/Loader.java
@@ -1,0 +1,3 @@
+package MrSSS.IO.Loader;
+
+public interface Loader {}
