+++ src/MrSSS/Item/ItemID.java
@@ -1,0 +1,6 @@
+package MrSSS.Item;
+
+public enum ItemID {
+    Bandana(),
+	Sword();
+}
