+++ src/MrSSS/Object/ObjectID.java
@@ -1,7 +1,6 @@
+package MrSSS.Object;
 
-
-public enum ID {
-
+public enum ObjectID {
     Player(),
     BasicEnemy(),
     Enemy(),
