+++ src/MrSSS/Skill/SkillList.java
@@ -1,5 +1,13 @@
+package MrSSS.Skill;
+
 import java.awt.Graphics;
 import java.awt.image.BufferedImage;
+
+import MrSSS.Game;
+import MrSSS.Object.Player;
+import MrSSS.Object.GameObject;
+import MrSSS.Object.ObjectID;
+import MrSSS.IO.MouseInput;
 
 public class SkillList {
     Player owner;
@@ -12,7 +20,7 @@ public class SkillList {
 	int y = mi.getY() - t.plyrToScrnY() + t.getY();
 	GameObject temp = GameObject.collision(Game.lvl.getHandler(),
 					       x, y, true);
-	if (temp != null && temp.getId() == ID.Obstacle){
+	if (temp != null && temp.getId() == ObjectID.Obstacle){
 	    System.out.println("Obstacle at " + Integer.toString(x) +
 			       ":" + Integer.toString(y) );
 	}
